+++ modules/cli/src/index.ts
@@ -10,7 +10,7 @@ import { Command } from "commander"
 import { CorsOptions } from "cors"
 import { config } from "dotenv"
 import { argv, cwd, exit } from "node:process"
-import { notifyListenerModules, startAnalysis, startServer } from "./App.js"
+import { notifyListenerModules, startAnalysis, startServer } from "./module/ModuleOrchestrator.js"
 import { createAnalysisCommand } from "./command/AnalysisCommand.js"
 import { createServerCommand } from "./command/ServerCommand.js"
 import { load, loadEnvironmentVariables } from "./module/ModuleLoader.js"
